+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 
 ### Fixed
 - Remote server running in W3C-protocol mode (eg. Selenium v3.5.3+) was erroneously detected as BrowserStack cloud service.
+- `--xdebug` option did not have any effect unless passed as the last option.
 
 ### Removed
 - `TestUtils` class which was already deprecated in 2.1.